front-end moved container components to layout
