+++ src/koekalenteri-frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Header from '../src/components/Header';
+import Header from '../src/layout/Header';
 import SearchPage from '../src/pages/SearchPage'
