+++ src/koekalenteri-frontend/src/layout/MainContainer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Container, Grid, Paper } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
-import EventText from './EventText';
+import EventText from '../components/EventText';
 
 const useStyles = makeStyles((theme) => ({
   container: {
