+++ src/koekalenteri-frontend/src/pages/SearchPage.tsx
@@ -1,8 +1,8 @@
 import { Container } from '@material-ui/core';
 import React from 'react';
-import InfoContainer from '../components/InfoContainer';
-import NotificationContainer from '../components/NotificationContainer';
-import MainContainer from '../components/MainContainer';
+import InfoContainer from '../layout/InfoContainer';
+import NotificationContainer from '../layout/NotificationContainer';
+import MainContainer from '../layout/MainContainer';
 
 const SearchPage = () => {
   return(
